+++ CHANGELOG.md
@@ -10,6 +10,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fix typos in export wins derived config
 
+## 2020-09-22
+
+### Added
+
+- HMRC Exporters matching pipeline
+
+### Changed
+
+- Support Array types in DSSGenericPipeline
+
 ## 2020-09-21
 
 ### Changed